refactor: auth 관련 컨트롤러 하나로 병합
